update nok logic to create revision record, update log to not show patient name
